Fix warnings and make this work.
Some notes: the WatchForFileDescriptor call was missing the flag saying that
we wanted to write, too. The error handling and ownership models are flakey at
best. If we lose our connection with the socket, we hang on shutdown. Note the
"XXX?" comment if we fail to send (I'm not sure what the right thing to do
there is). This fixes a nasty race in StartRil (it used to delete aMessage!).