Finish addressing bent's review comments
With this patch, the RILEvent is no longer a real event but simply a function
call on the worker's global object. This is much simpler and written this way,
we don't have to give dom/workers any knowledge of RIL at all.

This also moves the outer event handlers into JS. This means that we'll be
dispatching DOM events from the JS and not the C++.